commit ada555198cd80127358e38d852840fe349420739
Author: bedalton <Dsbadal@gmail.com>
Date:   Sat Aug 6 10:27:28 2022 -0700

    Add album icon

commit 96f6f5d77e2cad8cb448c46c102232b6ed07c44c
Author: bedalton <Dsbadal@gmail.com>
Date:   Sat Aug 6 10:27:19 2022 -0700

    Update project files

commit 78dec680e4cb65939cb42c255cc3bafb069b531a
Author: bedalton <Dsbadal@gmail.com>
Date:   Sat Aug 6 10:26:46 2022 -0700

    Update Agent-Util to comply with changes in V0.04 package changes

commit 947e39ff79f25dfe1c5de034776bf8228795d49b
Author: bedalton <Dsbadal@gmail.com>
Date:   Sat Aug 6 10:25:02 2022 -0700

    Add additional SPR formats

+++ src/main/java/com/badahori/creatures/plugins/intellij/agenteering/bundles/cobs/decompiler/readC2Cob.kt
@@ -55,7 +55,7 @@ private fun ByteStreamReader.readC2AgentBlock() : CobBlock.AgentBlock {
     val thumbWidth = uInt16
     val thumbHeight = uInt16
     val image = if (thumbWidth > 0 && thumbHeight > 0)
-        S16SpriteFile.parseFrame(this, position(), thumbWidth, thumbHeight, ColorEncoding.X_565, transparentBlack).toAwt()
+        S16SpriteFile.parseFrame(this, position().toLong(), thumbWidth, thumbHeight, ColorEncoding.X_565, transparentBlack).toAwt()
     else
         null
     skip(thumbWidth * thumbHeight * 2)
